working unit tests
