+++ tests/python/requirements.txt
@@ -1,1 +1,2 @@
+pandas==0.23.4
 pytest==5.4.3
