--- tests/python/unit/test_api_formatting.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-# This is a test file intended to be used with pytest
-# pytest automatically runs all the function starting with "test_"
-# see https://docs.pytest.org for more information
